+++ src/compas_occ/__init__.py
@@ -24,7 +24,7 @@ __author__ = ["tom van mele"]
 __copyright__ = "Block Research Group - ETH Zurich"
 __license__ = "MIT License"
 __email__ = "van.mele@arch.ethz.ch"
-__version__ = "0.5.0"
+__version__ = "0.6.0"
 
 
 HERE = os.path.dirname(__file__)
@@ -38,6 +38,6 @@ TEMP = os.path.abspath(os.path.join(HOME, "temp"))
 __all__ = ["HOME", "DATA", "DOCS", "TEMP"]
 
 __all_plugins__ = [
-    'compas_occ.geometry.curves',
-    'compas_occ.geometry.surfaces',
+    "compas_occ.geometry.curves",
+    "compas_occ.geometry.surfaces",
 ]